# command_center/backend/openapi.json
{
  "components": {
    "schemas": {
      "ActionMarker": {
        "properties": {
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "t": {
            "title": "T",
            "type": "integer"
          },
          "value": {
            "title": "Value",
            "type": "number"
          }
        },
        "required": [
          "t",
          "kind",
          "value"
        ],
        "title": "ActionMarker",
        "type": "object"
      },
...
      "CatalogResponse": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "empty": {
            "title": "Empty",
            "type": "boolean"
          },
          "hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hint"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/DatasetMetaOut"
            },
            "title": "Items",
            "type": "array"
          },
          "universe": {
            "items": {
              "type": "string"
            },
            "title": "Universe",
            "type": "array"
          }
        },
        "required": [
          "items",
          "count",
          "universe",
          "empty"
        ],
        "title": "CatalogResponse",
        "type": "object"
      },
      "CorporateActionsResponse": {
        "properties": {
          "adj_factor": {
            "items": {
              "$ref": "#/components/schemas/Point"
            },
            "title": "Adj Factor",
            "type": "array"
          },
          "dividends": {
            "items": {
              "$ref": "#/components/schemas/ActionMarker"
            },
            "title": "Dividends",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/SeriesMeta"
          },
          "splits": {
            "items": {
              "$ref": "#/components/schemas/ActionMarker"
            },
            "title": "Splits",
            "type": "array"
          },
          "ticker": {
            "title": "Ticker",
            "type": "string"
          }
        },
        "required": [
          "ticker",
          "adj_factor",
          "dividends",
          "splits",
          "meta"
        ],
        "title": "CorporateActionsResponse",
        "type": "object"
      },
...
      "CoverageOut": {
        "properties": {
          "end": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End"
          },
          "interval": {
            "title": "Interval",
            "type": "string"
          },
          "last_session_gap": {
            "title": "Last Session Gap",
            "type": "integer"
          },
          "missing_pct": {
            "title": "Missing Pct",
            "type": "number"
          },
          "n_expected": {
            "title": "N Expected",
            "type": "integer"
          },
          "n_present": {
            "title": "N Present",
            "type": "integer"
          },
          "start": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start"
          },
          "ticker": {
            "title": "Ticker",
            "type": "string"
          }
        },
        "required": [
          "ticker",
          "interval",
          "start",
          "end",
          "n_present",
          "n_expected",
          "missing_pct",
          "last_session_gap"
        ],
        "title": "CoverageOut",
        "type": "object"
      },
      "DatasetMetaOut": {
        "properties": {
          "end": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End"
          },
          "fetched_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fetched At"
          },
          "interval": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interval"
          },
          "modified_at": {
            "title": "Modified At",
            "type": "string"
          },
          "n_bytes": {
            "title": "N Bytes",
            "type": "integer"
          },
          "n_rows": {
            "title": "N Rows",
            "type": "integer"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "start": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start"
          },
          "tickers": {
            "items": {
              "type": "string"
            },
            "title": "Tickers",
            "type": "array"
          }
        },
        "required": [
          "source",
          "tickers",
          "interval",
          "n_rows",
          "n_bytes",
          "start",
          "end",
          "modified_at"
        ],
        "title": "DatasetMetaOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OHLCVBar": {
        "properties": {
          "adj": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Adj"
          },
          "c": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "C"
          },
          "h": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "H"
          },
          "low": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Low"
          },
          "o": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "O"
          },
          "t": {
            "title": "T",
            "type": "integer"
          },
          "v": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "V"
          }
        },
        "required": [
          "t",
          "o",
          "h",
          "low",
          "c",
          "adj",
          "v"
        ],
        "title": "OHLCVBar",
        "type": "object"
      },
      "OHLCVResponse": {
        "properties": {
          "bars": {
            "items": {
              "$ref": "#/components/schemas/OHLCVBar"
            },
            "title": "Bars",
            "type": "array"
          },
          "interval": {
            "title": "Interval",
            "type": "string"
          },
          "meta": {
            "$ref": "#/components/schemas/SeriesMeta"
          },
          "ticker": {
            "title": "Ticker",
            "type": "string"
          }
        },
        "required": [
          "ticker",
          "interval",
          "bars",
          "meta"
        ],
        "title": "OHLCVResponse",
        "type": "object"
      },
...
      "Point": {
        "properties": {
          "t": {
            "title": "T",
            "type": "integer"
          },
          "v": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "V"
          }
        },
        "required": [
          "t",
          "v"
        ],
        "title": "Point",
        "type": "object"
      },
...
      "RefreshRequest": {
        "properties": {
          "tickers": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tickers"
          }
        },
        "title": "RefreshRequest",
        "type": "object"
      },
      "RefreshResponse": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "rows": {
            "title": "Rows",
            "type": "integer"
          },
          "tickers": {
            "items": {
              "type": "string"
            },
            "title": "Tickers",
            "type": "array"
          }
        },
        "required": [
          "ok",
          "tickers",
          "rows",
          "message"
        ],
        "title": "RefreshResponse",
        "type": "object"
      },
...
      "SeriesMeta": {
        "description": "Honest provenance/shape: where it came from, how big, whether it was downsampled.",
        "properties": {
          "downsampled": {
            "default": false,
            "title": "Downsampled",
            "type": "boolean"
          },
          "empty": {
            "default": false,
            "title": "Empty",
            "type": "boolean"
          },
          "end": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End"
          },
          "n_raw": {
            "default": 0,
            "title": "N Raw",
            "type": "integer"
          },
          "n_returned": {
            "default": 0,
            "title": "N Returned",
            "type": "integer"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "start": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start"
          }
        },
        "title": "SeriesMeta",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Bloomberg-style research terminal API over quant_core.",
    "title": "command_center",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/catalog": {
      "get": {
        "operationId": "getCatalog",
        "parameters": [
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Catalog Route"
      }
    },
    "/api/v1/catalog/{ticker}/coverage": {
      "get": {
        "operationId": "getCoverage",
        "parameters": [
          {
            "in": "path",
            "name": "ticker",
            "required": true,
            "schema": {
              "pattern": "^[A-Z0-9.\\-]{1,15}$",
              "title": "Ticker",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "interval",
            "required": false,
            "schema": {
              "default": "1d",
              "title": "Interval",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "calendar",
            "required": false,
            "schema": {
              "default": "NYSE",
              "title": "Calendar",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CoverageOut"
                  },
                  "title": "Response Getcoverage",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Coverage Route"
      }
    },
    "/api/v1/corporate-actions/{ticker}": {
      "get": {
        "operationId": "getCorporateActions",
        "parameters": [
          {
            "in": "path",
            "name": "ticker",
            "required": true,
            "schema": {
              "pattern": "^[A-Z0-9.\\-]{1,15}$",
              "title": "Ticker",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "in": "query",
            "name": "end",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          },
          {
            "in": "query",
            "name": "interval",
            "required": false,
            "schema": {
              "default": "1d",
              "title": "Interval",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CorporateActionsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Corporate Actions"
      }
    },
...
    "/api/v1/ohlcv/{ticker}": {
      "get": {
        "operationId": "getOhlcv",
        "parameters": [
          {
            "in": "path",
            "name": "ticker",
            "required": true,
            "schema": {
              "pattern": "^[A-Z0-9.\\-]{1,15}$",
              "title": "Ticker",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "in": "query",
            "name": "end",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          },
          {
            "in": "query",
            "name": "interval",
            "required": false,
            "schema": {
              "default": "1d",
              "title": "Interval",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "in": "query",
            "name": "max_points",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Points"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OHLCVResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Ohlcv"
      }
    },
...
    "/api/v1/refresh": {
      "post": {
        "description": "Best-effort LOCAL fetch into the cache via quant_core (never used in CI; yfinance is best-effort).",
        "operationId": "refresh",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/RefreshRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh Route"
      }
    },
...
  }
}